refactor: use component entity builder

+++ app/src/main/java/com/example/crazyball/model/tables/entities/ComponentEntityBuilder.java
@@ -4,13 +4,13 @@ import com.example.crazyball.model.EComponentType;
 
 public class ComponentEntityBuilder {
 
-    private final int levelId;
+    private final long levelId;
     private int locationX;
     private int locationY;
     private String type;
     private int imageId;
 
-    public ComponentEntityBuilder(int levelId) {
+    public ComponentEntityBuilder(long levelId) {
         this.levelId = levelId;
     }
 
@@ -34,7 +34,7 @@ public class ComponentEntityBuilder {
         return this;
     }
 
-    public int getLevelId() {
+    public long getLevelId() {
         return levelId;
     }
